+++ src/engine/engine_vis_init.c
@@ -88,7 +88,8 @@ const char* mjVISSTRING[mjNVISFLAG][3] = {
   {"Center of &Mass", "0", "M"},
   {"S&elect Point",   "0", "E"},
   {"Static Bo&dy",    "1", "D"},
-  {"Skin",            "1", ";"}
+  {"Skin",            "1", ";"},
+  {"Body Tree",       "0", "`"}
 };